requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
